refactor: rename file according to new convention. Move update function to UpdateUtil to match Configuration test setup
