--- 03_Modules/Transactions/test/providers/update.ts
@@ -1,1 +1,0 @@
-export type UpdateFunction<T> = (site: T) => void;
