+++ 03_Modules/Transactions/test/utils/UpdateUtil.ts
@@ -1,0 +1,11 @@
+export type UpdateFunction<T> = (item: T) => void;
+
+export const applyUpdateFunction = <T>(
+  item: T,
+  updateFunction?: UpdateFunction<T>,
+): T => {
+  if (updateFunction) {
+    updateFunction(item);
+  }
+  return item;
+};
